Refactored transferring data to mailjet and updating on change of users table
